# Migrate build process from Webpack to Rspack for faster builds