--- a/frontend/package.json
+++ b/frontend/package.json
@@ -86,8 +86,7 @@
     "prebuild": "del-cli ./dist",
     "build": "rspack build --config rspack.prod.js",
     "build-dev": "rspack build --mode development",
-    "build:analyze": "BUNDLE_ANALYZER=true rspack --config rspack.prod.js",
-    "build:doctor": "RSDOCTOR=true rspack --config rspack.prod.js",
+    "build:analyze": "RSDOCTOR=true rspack --config rspack.prod.js",
     "start": "rspack serve --mode=development --config rspack.dev.js",
     "serve": "node scripts/serve.js",
     "lint": "eslint --fix .",
@@ -103,7 +102,6 @@
     "@rsdoctor/rspack-plugin": "^0.4.12",
     "@rspack/cli": "^1.1.8",
     "@rspack/core": "^1.1.8",
-    "@types/webpack-bundle-analyzer": "^4.7.0",
     "@web/dev-server-esbuild": "^0.3.3",
     "@web/dev-server-import-maps": "^0.2.0",
     "@web/dev-server-rollup": "^0.6.1",
@@ -114,8 +112,7 @@
     "rollup-plugin-typescript-paths": "^1.4.0",
     "sinon": "^12.0.1",
     "ts-checker-rspack-plugin": "^1.1.0",
-    "ts-lit-plugin": "^2.0.1",
-    "webpack-bundle-analyzer": "^4.10.1"
+    "ts-lit-plugin": "^2.0.1"
   },
   "optionalDependencies": {
     "@esm-bundle/chai": "^4.3.4-fix.0",