--- a/frontend/package.json
+++ b/frontend/package.json
@@ -77,7 +77,6 @@
     "style-loader": "^3.3.0",
     "tailwindcss": "^3.4.1",
     "ts-checker-rspack-plugin": "^1.1.0",
-    "ts-lit-plugin": "^2.0.2",
     "typescript": "^5.3.3",
     "update-dotenv": "^1.1.1",
     "url-pattern": "^1.0.3",
@@ -112,7 +111,8 @@
     "lint-staged": "^13.1.0",
     "prettier-plugin-tailwindcss": "^0.5.12",
     "rollup-plugin-typescript-paths": "^1.4.0",
-    "sinon": "^12.0.1"
+    "sinon": "^12.0.1",
+    "ts-lit-plugin": "^2.0.2"
   },
   "optionalDependencies": {
     "@esm-bundle/chai": "^4.3.4-fix.0",