Swaps out Webpack for [Rspack](https://rspack.dev/), a rust-based bundler intended to be a mostly-drop-in replacement for Webpack with much better performance.

## Correctness

I've tested the following, but would appreciate more testing!

It correctly:
- [x] builds and serves the app for development when `yarn start` is run, without errors
- [x] rebuilds and reloads the page when files are changed
- [x] surfaces lint & type errors during development as expected
- [x] builds the app with `yarn build`
- [x] analyses the build output with `yarn build:analyze` as expected
- [x] runs the `serve` script, extending from the dev config, with `yarn serve`

## Benchmarks

Run with `hyperfine --show-output --shell zsh --warmup 3 'yarn build'` on my 2021 M1 Pro MacBook Pro in Ghostty.

Webpack (`main` @ cf1d97ab687ab8ebe8f65c9f114f383aca927717)

```
  Time (mean ± σ):     15.674 s ±  0.317 s    [User: 58.462 s, System: 3.815 s]
  Range (min … max):   15.147 s … 16.246 s    10 runs
```

Rspack (`frontend-rspack-experiment`@ 2726ed3)
```
  Time (mean ± σ):      6.672 s ±  0.055 s    [User: 18.773 s, System: 2.704 s]
  Range (min … max):    6.556 s …  6.759 s    10 runs
```

That's a 2.349× speedup on average!

## Additional features

- I've replaced `webpack-bundle-analyser` with [Rsdoctor](https://rsdoctor.dev/), which, in addition to showing the same bundle size analysis, also shows a bunch of other diagnostics, suggestions, errors, logs, and other useful bits of info.

## Caveats

Users will need to re-run `yarn install` (or just `yarn`) once this is merged.

**Supply chain security**: Rspack is a project from ByteDance, which the US Government seems have some feelings about. The software itself is MIT-licensed, so I can't imagine there's going to be a huge issue here, but maybe this is something we want to be aware of?