@SuaYoo 2025-01-07
Very nice, benchmark results on my 2021 Macbook Pro, M1 Pro, macOS 14.5.

```
Rspack, Node v20.17.0
  Time (mean ± σ):      8.215 s ±  0.096 s    [User: 23.400 s, System: 2.443 s]
  Range (min … max):    8.038 s …  8.350 s    10 runs

Rspack, Node v20.17.0
  Time (mean ± σ):      7.802 s ±  0.112 s    [User: 22.225 s, System: 2.557 s]
  Range (min … max):    7.689 s …  8.004 s    10 runs

Webpack, Node v22.11.0
  Time (mean ± σ):     19.413 s ±  0.305 s    [User: 65.024 s, System: 3.754 s]
  Range (min … max):   18.896 s … 20.027 s    10 runs

Webpack, Node v20.17.0
  Time (mean ± σ):     20.513 s ±  0.925 s    [User: 71.152 s, System: 3.736 s]
  Range (min … max):   19.593 s … 22.406 s    10 runs
```

Regarding ByteDance, there's definitely **strong** feelings. It does look like some of webpack's core maintainers are on the Rspack team. Will defer to @ikreymer as product owner.